requirements.txt
python-dotenv
azure-cognitiveservices-vision-face==0.6.0
msrest
pillow
matplotlib
